+++ Content.Shared/Chemistry/Components/SolutionManager/ExaminableSolutionComponent.cs
@@ -1,14 +1,27 @@
 ﻿namespace Content.Shared.Chemistry.Components.SolutionManager;
 
+/// <summary>
+///     Component for examining a solution with shift click.
+/// </summary>
 [RegisterComponent]
 public sealed partial class ExaminableSolutionComponent : Component
 {
-    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    /// <summary>
+    ///     The solution being examined.
+    /// </summary>
+    [DataField]
     public string Solution = "default";
 
     /// <summary>
-    /// If false then the hidden solution is always visible.
+    ///     If true, the solution must be held to be examined.
     /// </summary>
     [DataField]
     public bool HeldOnly;
+
+    /// <summary>
+    ///     If true, the examine text will give an approximation of the remaining solution.
+    ///     If false, the exact unit count will be shown.
+    /// </summary>
+    [DataField]
+    public bool VagueExamine;
 }
